backend/requirements.txt
# FinancIA 2030 - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pgvector==0.2.3

# Search & Vector Store
opensearch-py==2.4.2
elasticsearch==8.11.0

# Cache & Message Queue
redis==5.0.1
kafka-python==2.0.2
aiokafka==0.9.0

# Object Storage
minio==7.2.0

# Document Processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0
openpyxl==3.1.2
pdf2image==1.16.3
opencv-python==4.8.1.78
camelot-py[cv]==0.11.0
# textract==1.6.5  # Removed due to invalid metadata in pip 25+
pdfplumber==0.10.3
python-pptx==0.6.23
langdetect==1.0.9
reportlab==4.0.7

# NLP & ML
spacy==3.6.1  # Last stable version compatible with Pydantic 2
transformers==4.35.2
huggingface_hub>=0.20.0,<0.22.0  # Pin compatible version range
sentence-transformers==2.7.0  # Updated for compatibility with newer huggingface_hub
# PyTorch with CUDA support (installed separately in Dockerfile.backend.gpu)
# torch>=2.0.0  # Latest stable version
# torchvision>=0.15.0
# torchaudio>=2.0.0
# LangChain & OpenAI
langchain>=0.0.350
langchain-openai>=0.0.2
openai>=1.6.1
tiktoken>=0.5.2
anthropic>=0.8.0
numpy>=1.26.0  # Updated for better compatibility
scikit-learn>=1.3.0

# Ontology & Semantic Web (Sprint 2 + 3)
rdflib==7.1.1
owlready2==0.45
SPARQLWrapper==2.0.0

# Audio/Video Processing
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1

# ML Ops
mlflow==2.9.2
evidently==0.4.33  # Updated to support pydantic 2.x

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-keycloak==3.7.0
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.26.0  # Updated for evidently compatibility
opentelemetry-sdk==1.26.0  # Matches semantic-conventions 0.47b0
opentelemetry-instrumentation-fastapi==0.47b0

# LLM Observability con Arize Phoenix  
arize-phoenix>=5.0.0  # Updated for Pydantic 2.x compatibility
openinference-instrumentation-openai==0.1.11
openinference-instrumentation-langchain==0.1.20
opentelemetry-exporter-otlp==1.26.0  # Matches SDK version

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
orjson==3.9.10
pyyaml==6.0.1
python-magic==0.4.27
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Development
black==23.12.0
ruff==0.1.7
mypy==1.7.1
pre-commit==3.6.0
